Get PWM working with TIM1 periph
In the code right now, the onboard LED should blink twice as slowly
as an LED hooked up to the D7 pin on the NUCLEO board. This is
because the D7 pin is PWM'd and an update event only occurs
after a cycle of the TIM1.

Now that we have the foundation, we can write the WS2812B code
to just update the CCR and ARR registers. For the low pulse, we can
just set the period to be the 300 us and the duty cycle to be 0. Then
if we are done we can simply freeze the counter / reset it.