Get PWM working? - needs to be tested
Using ARR and CCR, it seems like PWM is working
Would be much easier to see with a logic analyzer whether or not
we are respecting the low and high times. Before sending this to a
NeoPixel strip I would like to verify that.

This is just a template/rough sketch of what I think this will
look like.