Add GPIO driver similiar to Arduino pinMode() and digitalWrite()
Goal is to simplify GPIO setup for the other drivers.
As part of the various peripheral init, GPIOs need to be setup. As of now,
I had just been doing the low-level register writes.

With this commit, we can successfully use 2 simple functions to
initialize a GPIO (returns a GPIO struct) and then set/reset the specific pin.

The GPIO struct encapsulates 2 pieces of information which is needed for
GPIO configuration and writing: the port (A,B,...,H) and the pin (1,2,...,15)

Using #define preprocessor macros, we can map these port and pin values
to the appropiate register bank which is transparent to the user.

This is slightly different to how STM32 has set up their LL drivers
and the standard peripheral library.