Change UART to use GPIO driver and fix GPIO bug
enums were using hexadecimal representation but I meant to use
binary for the gpio_mode_enum. For the others I meant to use
decimal, so I changed the values respectively.

Also add a change to GPIO so that the actual base register
is saved in the struct. I thought this would improve efficiency
since we don't need to calculate the address of the base address anymore.
Stepping through dissasembly I'm not too sure, but meh. Isn't a big deal
to do and it simplifies the GPIO driver code.

UART code works as expected with GPIO init.